airflow/requirements.txt
pandas>2
psycopg2-binary==2.9.10
word2number==1.1 
dbt-postgres